Adding support for an external TUI
An external TUI (cryptic-tui) written in Rust is
now able to communicate with the Cryptic client
node as a hidden node.

The module: cryptic_rpc.erl is inoked by the
external TUI node to perform actions.

The module: cryptic_tui_bridge.erl is a gen_server
that subscribes to bus events which are relayed
to the external TUI node.